+++ src/pc-v5/constants.js
@@ -8,13 +8,8 @@ const packageJSON = require('./package.json')
 
 const PLATFORM = os.platform()
 const ZADARK_VERSION = packageJSON.version
-const DARK_TYPE_LABEL = {
-  dark: 'default',
-  dark_dimmed: 'dimmed'
-}
 
 module.exports = {
   PLATFORM,
-  ZADARK_VERSION,
-  DARK_TYPE_LABEL
+  ZADARK_VERSION
 }
