+++ src/pc/constants.js
@@ -8,8 +8,13 @@ const packageJSON = require('./package.json')
 
 const PLATFORM = os.platform()
 const ZADARK_VERSION = packageJSON.version
+const DARK_TYPE_LABEL = {
+  dark: 'default',
+  dark_dimmed: 'dimmed'
+}
 
 module.exports = {
   PLATFORM,
-  ZADARK_VERSION
+  ZADARK_VERSION,
+  DARK_TYPE_LABEL
 }
